refactor : ResultType Finished 내부로 이동 및 State 패키지 간소화
