+++ src/main/java/blackjack/domain/participant/BlackJackParticipant.java
@@ -2,8 +2,8 @@ package blackjack.domain.participant;
 
 import blackjack.domain.Hand;
 import blackjack.domain.card.Card;
+import blackjack.domain.state.NotStarted;
 import blackjack.domain.state.State;
-import blackjack.domain.state.StateFactory;
 
 import java.util.ArrayList;
 import java.util.Objects;
@@ -17,7 +17,7 @@ public abstract class BlackJackParticipant {
     public BlackJackParticipant(String name) {
         this.hand = new Hand(new ArrayList<>());
         this.name = new Name(name);
-        this.state = StateFactory.getInstance();
+        this.state = new NotStarted();
     }
 
     abstract public void draw(Card card);
