+++ src/test/java/blackjack/domain/PlayerTest.java
@@ -6,6 +6,9 @@ import blackjack.domain.card.Shape;
 import blackjack.domain.participant.Dealer;
 import blackjack.domain.participant.Player;
 import blackjack.domain.state.*;
+import blackjack.domain.state.finished.BlackJackState;
+import blackjack.domain.state.finished.Bust;
+import blackjack.domain.state.finished.Stay;
 import blackjack.exception.InvalidNameInputException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
