+++ src/test/java/blackjack/domain/state/BlackJackStateTest.java
@@ -1,12 +1,11 @@
 package blackjack.domain.state;
 
 import blackjack.domain.TestSetUp;
-import blackjack.domain.participant.Dealer;
+import blackjack.domain.state.finished.BlackJackState;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class BlackJackStateTest {
     @Test
